refactor: update imports to use client package for KeyGenerator and APIKeyGenerator
