+++ profanity-domain/src/test/java/app/application/apikey/ClientsCommandServiceTest.java
@@ -1,6 +1,8 @@
 package app.application.apikey;
 
+import app.application.client.APIKeyGenerator;
 import app.application.client.ClientsCommandService;
+import app.application.client.KeyGenerator;
 import app.core.exception.BusinessException;
 import app.domain.InmemoryClientsRepository;
 import app.dto.request.ClientRegistCommand;
