+++ profanity-domain/src/test/java/app/application/apikey/KeyGeneratorTest.java
@@ -1,5 +1,7 @@
 package app.application.apikey;
 
+import app.application.client.APIKeyGenerator;
+import app.application.client.KeyGenerator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
